master

+++ public/js/student/doHomework.js
@@ -60,7 +60,6 @@ $(function(){
         $(".hw-order-index>span:nth-child("+num+"),.answer-sheet .hw-order>span:nth-child("+num+")").toggleClass("query_order");
     });
 
-
     //设置题目ul的宽度
     var li_width = 918;
     var li_num = $(".do-hw .exercise-box>li").length;
@@ -95,12 +94,14 @@ $(function(){
         left = parseInt($(".do-hw .exercise-box").css("left")) - li_width;
         $(".do-hw .exercise-box").css("left",left);
 
-
+        //右上角题号更换
+        //if(left !== -ul_width+li_width) {
+        //    $(".big-num").text(parseInt($(".big-num").text()) + 1);
+        //}
         $(".big-num").text(parseInt($(".big-num").text()) + 1);
 
         isEnd(left);
     });
-
 
     //右移
     $("#fa-angle-left").click(function(){
@@ -110,6 +111,10 @@ $(function(){
         },300);
 
         single_time();
+
+        //if(left !== -ul_width+li_width) {
+        //    $(".big-num").text(parseInt($(".big-num").text()) - 1);
+        //}
 
         left = left + li_width;
         $(".do-hw .exercise-box").css("left",left);
@@ -151,7 +156,7 @@ $(function(){
         }else {
             getOrderAndBlue();
         }
-    
+    });
 
     //禁用Tab键和回车键
     $(window).keydown(function(event){
@@ -159,6 +164,7 @@ $(function(){
             event.preventDefault();
         }
     });
+
 
     //排序题
     $(".sortable").sortable();
@@ -172,19 +178,6 @@ $(function(){
     $("#handPaper").click(function(){
         var work_id = $("#work_id").attr("value");
         clearInterval(timer);
-
-        //var total = [
-        //    {id:"1", answer:"C", last:0}, //单选题
-        //    {id:"2", answer:"BC", last:0}, //多选题
-        //    {id:"3", answer:["111","222"], last:0}, //填空题
-        //    {id:"4", answer:"错误", last:0}, //判断题
-        //    {id:"5", answer:"BDAC", last:0}, //排序题
-        //    {id:"6", answer:["data-...","data-..."], last:0}, //画图题
-        //    {id:"7", answer:["2连3","1连2","3连1"], last:0}, //连线题
-        //    {id:"8", answer:{"text":"aaa"}, last:0}, //简答题
-        //    {id:"8", answer:[{"id":"101", "answer":"B"},{},{}], last:0}, //听力题
-        //];
-
         var total = [];
         total.total_time = parseInt((new Date().getTime() - sessionStorage.getItem("ic_start_time"))/1000);
 
@@ -193,67 +186,75 @@ $(function(){
             //保存单题的答案
             var obj = {
                 "id": $(item).attr("data-id"),
-                "answer": "",
-                "last": 0
+                "answer": [],
+                "parent_id":"",
+                "last": 0,
             };
 
-
-			var arr = []; //单题
-			var arr_big = []; //多题
-			var dom = "";
+            var arr = []; //单题
+            var arr_big = []; //多题
+            var dom = "";
             //保存图文并存的答案
             var img_text = {};
 
             if(type === "单选题"){
                 obj.answer = $(item).find(".ic-radio.active input").val() == null ? "" : $(item).find(".ic-radio.active input").val();
+                obj.parent_id =  $(item).find(".ic-blue .do-hw-type").attr('parent-id');
             }else if(type === "多选题"){
                 $(item).find(".radio-wrap .ic-radio.active input").each(function(i,n){
-                    obj.answer += $(n).val()+',';
+                   obj.answer.push(Number($(n).val()));
                 });
-                 obj.answer= obj.answer.substring(0,obj.answer.length-1)
+                console.log(obj.answer)
+                if (obj.answer[0] == null) {
+                    obj.answer = "";
+                }
+                obj.parent_id =  $(item).find(".ic-blue .do-hw-type").attr('parent-id')
             }else if(type === "填空题" || type === "多空题"){
-            	$(item).find(".blank-item").each(function(i,n){
+                $(item).find(".blank-item").each(function(i,n){
                     arr.push($(n).text());
                 });
                 obj.answer = arr;
+                obj.parent_id =  $(item).find(".ic-blue .do-hw-type").attr('parent-id')
             }else if(type === "判断题"){
-            	dom = $(item).find(".answer-box .pan-duan");
-            	if(dom.hasClass("rightActive")){
-            		obj.answer = "1";
-            	}else if(dom.hasClass("wrongActive")){
-            		obj.answer = "0";
-            	}else {
-            		obj.answer = "";
-            	}
+                dom = $(item).find(".answer-box .pan-duan");
+                if(dom.hasClass("rightActive")){
+                    obj.answer = "1";
+                }else if(dom.hasClass("wrongActive")){
+                    obj.answer = "0";
+                }else {
+                    obj.answer = "";
+                }
+                obj.parent_id =  $(item).find(".ic-blue .do-hw-type").attr('parent-id')
             }else if(type === "完形填空"){
-            	$(item).find(".wan-xing-tk-option").each(function(i,n){
-            		arr.push($(n).find(".ic-radio.active input").val());
-            	});
-            	obj.answer = arr;
-            }else if(type === "排序题"){
-            	$(item).find(".exer-list-ul>li>span").each(function(i,n){
-                    obj.answer += $(n).text().slice(2,3)+',';
+                $(item).find(".wan-xing-tk-option").each(function(i,n){
+                    arr.push($(n).find(".ic-radio.active input").val());
                 });
-                obj.answer= obj.answer.substring(0,obj.answer.length-1)
+                obj.answer = arr;
+            }else if(type === "排序题"){
+                $(item).find(".exer-list-ul>li>span").each(function(i,n){
+                    obj.answer.push($(n).attr('exercise-id'));
+                });
+                obj.parent_id =  $(item).find(".ic-blue .do-hw-type").attr('parent-id')
             }else if(type === "画图题" || type === "作文题" || type === "计算题"){
-            	$(item).find(".one-img>img").each(function(i,n){
-            		arr.push($(n).attr("src"));
-            	});
-            	obj.answer = arr;
+                $(item).find(".one-img>img").each(function(i,n){
+                    arr.push($(n).attr("src"));
+                });
+                obj.answer = arr;
             }else if(type === "连线题"){
-            	obj.answer = JSON.parse(sessionStorage.getItem("ic_lianXianTi"+$(item).attr("data-id")));
+                obj.answer = JSON.parse(sessionStorage.getItem("ic_lianXianTi"+$(item).attr("data-id")));
+                obj.parent_id =  $(item).find(".ic-blue .do-hw-type").attr('parent-id')
             }else if(type === "简答题"){
-            	if($(item).find(".editor-content img").length !== 0){
-            		$(item).find(".editor-content img").each(function(i,n){
-            			arr.push($(n).attr("src"));
-            		});
+                if($(item).find(".editor-content img").length !== 0){
+                    $(item).find(".editor-content img").each(function(i,n){
+                        arr.push($(n).attr("src"));
+                    });
                     img_text.img = arr;
-            	}else {
+                }else {
                     img_text.text = $(item).find(".editor-content").text();
-            	}
+                }
                 obj.answer = img_text;
             }else if(type === "听力题" || type === "阅读题" || type === "解答题"){
-            	$(item).find(".one-hw").each(function(i,item){
+                $(item).find(".one-hw").each(function(i,item){
                     var obj_child = {
                         "id": $(item).attr("data-id"),
                         "answer": ""
@@ -273,7 +274,7 @@ $(function(){
                         $(item).find(".blank-item").each(function(i,n){
                             arr.push($(n).text());
                         });
-                        obj_child.answer = arr;
+                        obj_child.answer = arr; 
                     }else if(type === "判断题"){
                         dom = $(item).find(".answer-box .pan-duan");
                         if(dom.hasClass("rightActive")){
@@ -308,33 +309,39 @@ $(function(){
                     }
                     arr_big.push(obj_child);
                 });
-
                 obj.answer = arr_big;
             }
-            
             total.push(obj);
         });
         var param = clearUp(total); //传给后台的作业答案参数
         param._token = token;
         param.work_id = work_id;
-        $.post("/homeworkScores",param,function(result){
-            var course = $("#course_id").attr('value');
-            if (result == 200 || result == 1) {   
-                window.location.href = "/learningCenter/" + course + "/1/3/" + work_id;
-            }
-
-            sessionStorage.clear();
-        });
-
+        console.log(param['data'][0]['parent_id']);
+        console.log(param)
+        //console.log(accuracy)
+        if (param['data'][0]['parent_id'] != ""){
+            $.post("/sameScore",param,function(result){
+                var course = $("#course_id").attr('value');
+                 window.location.href = "/learningCenter/" + course + "/1/6/" + work_id +"/"+ accuracy;
+            });
+        }else{
+            $.post("/homeworkScores",param,function(result){
+                var course = $("#course_id").attr('value');
+                if (result == 200 || result == 1) {
+                   window.location.href = "/learningCenter/" + course + "/1/3/" + work_id;
+                }
+                sessionStorage.clear();
+            });
+        }
         //把单题时间整合到total中,li_num
         var store = window.sessionStorage;
         for(var key in store){
             if(Number(key.slice(3))){
+                console.log(key);
                 total[Number(key.slice(3))-1].last = JSON.parse(store[key]).last;
             }
         }
     });
-
 
     //整理arr类型为obj
     function clearUp(arr){
@@ -350,103 +357,8 @@ $(function(){
                 obj.data.push(arr[key]);
             }
         }
-
         return obj;
     }
 });
-
-/*    //毫秒格式化为00:00
-    function changeTime(millisecond){
-        //计算"分：秒"
-        var m = parseInt(millisecond/1000/60);
-        var s = parseInt(millisecond/1000 - m*60);
-        if(m<10){
-            m = "0" + m;
-        }
-        if(s<10){
-            s = "0" + s;
-        }
-        return m + ":" + s;
-    }
-
-
-    function isEnd(left){
-        if(left === -ul_width + li_width){
-            $(".fa-angle-right,.hw-order-index,.exer-num").hide();
-            $(".fa-angle-left,.answer-sheet-icon").show();
-        }else if(left === 0){
-            $(".fa-angle-right,.hw-order-index,.exer-num").show();
-            $(".fa-angle-left,.answer-sheet-icon").hide();
-        }else {
-            $(".fa-angle-right,.fa-angle-left,.hw-order-index,.exer-num").show();
-            $(".answer-sheet-icon").hide();
-        }
-    }
-
-
-    function single_time(){
-        var current = new Date().getTime();
-        obj_time.id = $(".big-num").text();
-        var last = Math.round((current - obj_time.start_time)/1000);
-        var ic_id = sessionStorage.getItem("ic_"+obj_time.id);
-        if(ic_id){
-            obj_time.last = JSON.parse(ic_id).last + last;
-        }else {
-            obj_time.last = last;
-        }
-        obj_time.start_time = current;
-        sessionStorage.setItem("ic_"+obj_time.id,JSON.stringify(obj_time));
-    }*/
-
-
-    //整理arr类型为obj
-/*    function clearUp(arr){
-        var obj = {
-            "data":[],
-            "total_time": ""
-        };
-
-        for(var key in arr){
-            if(key==="total_time"){
-                obj.total_time = arr[key];
-            }else {
-                obj.data.push(arr[key]);
-            }
-        }
-        return obj;
-    }
-});
-
-*/
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
